+++ src/WeihanLi.Common.Logging.Log4Net/ElasticSearchAppender.cs
@@ -1,10 +1,10 @@
-﻿using System;
+﻿using log4net.Appender;
+using log4net.Core;
+using System;
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Text;
 using System.Threading;
-using log4net.Appender;
-using log4net.Core;
 using WeihanLi.Common.Helpers;
 using WeihanLi.Extensions;