+++ src/WeihanLi.Common/DependencyInjection/ServiceContainerModule.cs
@@ -1,6 +1,11 @@
 ﻿namespace WeihanLi.Common.DependencyInjection
 {
-    public abstract class ServiceContainerModule
+    public interface IServiceContainerModule
+    {
+        void ConfigureServices(IServiceContainerBuilder serviceContainerBuilder);
+    }
+
+    public abstract class ServiceContainerModule : IServiceContainerModule
     {
         public abstract void ConfigureServices(IServiceContainerBuilder serviceContainerBuilder);
     }