+++ src/WeihanLi.Common/Logging/LogHelperLoggerLevel.cs
@@ -1,9 +1,9 @@
 ﻿namespace WeihanLi.Common.Logging
 {
     /// <summary>
-    /// LoggerLevel
+    /// LogLevel
     /// </summary>
-    public enum LogHelperLevel
+    public enum LogHelperLogLevel
     {
         /// <summary>
         /// All logging levels