+++ src/WeihanLi.Common/Logging/LogHelperProvider.cs
@@ -2,11 +2,6 @@
 {
     public interface ILogHelperProvider
     {
-        ILogHelperLogger CreateLogger(string categoryName);
-    }
-
-    internal class NullLogHelperProvider : ILogHelperProvider
-    {
-        public ILogHelperLogger CreateLogger(string categoryName) => NullLogHelperLogger.Instance;
+        void Log(LogHelperLoggingEvent loggingEvent);
     }
 }