+++ docs/ReleaseNotes.md
@@ -4,6 +4,21 @@
 
 .netstandard2.0 based also support for netfx4.5 common helpers and extensions
 
+
+### [WeihanLi.Common 1.0.28](https://www.nuget.org/packages/WeihanLi.Common/1.0.28)
+
+- refact logging
+- update `JsonResultModel`/`JsonResultStatus` => `ResultModel`/`ResultStatus`
+- add `NetHelper.IsPrivateIP`/`IPNetwork`
+- add httpHeader parameters for `HttpHelper.HttpPostFile`
+
+### [WeihanLi.Common 1.0.26](https://www.nuget.org/packages/WeihanLi.Common/1.0.26)
+
+- update event, add async support for publish and subscribe/unsubscribe
+- update di, add `ServiceContainerBuilder`/`ServiceContainerModule` to register service
+- update cron, export `CronExpression`
+- add `ArrayHelper.Empty` for net45,update `ConfigurationHelper` for netstandard2.0
+
 ### [WeihanLi.Common 1.0.25](https://www.nuget.org/packages/WeihanLi.Common/1.0.25)
 
 - add `ValidateResultModel`
@@ -98,6 +113,10 @@
 - update repository, add support for columns mapping
 
 ## WeihanLi.Common.Logging.Log4Net
+
+### [WeihanLi.Common.Logging.Log4Net 1.0.28](https://www.nuget.org/packages/WeihanLi.Common.Logging.Log4Net/1.0.28)
+
+- refact logging
 
 ### [WeihanLi.Common.Logging.Log4Net 1.0.23.8](https://www.nuget.org/packages/WeihanLi.Common.Logging.Log4Net/1.0.23.8)